[컴포넌트] meet-up-detail 뷰, 로직 커스텀 훅으로 분리
