+++ src/pages/meet-up-detail/meet-up-detail-templete.js
@@ -1,11 +1,11 @@
-import axios from 'axios'
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import styled from 'styled-components'
 import FooterBtn from '../../components/atoms/buttons/footer-button'
 import MeetUpPost from '../../components/organisms/meet-up/meet-up-post'
 import BottomPopUp from '../../components/organisms/modal/bottom-pop-up'
 import PopUp from '../../components/organisms/modal/pop-up'
 import useMeetUpBottom from '../../hooks/useMeetUpBottom'
+import useMeetUpDetail from './useMeetUpDetail'
 
 const Main = styled.main`
     width: 100%;
@@ -17,7 +17,7 @@ const Main = styled.main`
     box-sizing: border-box;
 `
 
-function MeetUpDetailTemp({ id, detail, status }) {
+function MeetUpDetailTemp({ detail, status }) {
     const [isOpenPopUP, setIsOpenPopUp] = useState(false)
     const [isBottomPopUP, setIsBottomPopUp] = useState(true)
     const popUpFunctions = useMeetUpBottom(
@@ -26,77 +26,11 @@ function MeetUpDetailTemp({ id, detail, status }) {
         status,
         detail
     )
-
-    const [{ footerText, popUpText, handleOk }, setInfo] = useState({
-        footerText: '',
-        popUpText: '',
-        handleOk: function () {},
-    })
-    const uri = 'http://175.45.195.94:9999/api'
-    const config = {
-        headers: {
-            Authorization:
-                'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzUxMiJ9.eyJzdWIiOiI2MzBmNTMwNGM2ODU3MTE5M2MxZDhiNzIiLCJleHAiOjE2NjUxMDQ5NTAsImlhdCI6MTY2NDkzMjE1MCwiZW1haWwiOiJndWVzdDIyMkBnbWFpbC5jb20ifQ.5AggSnpnpTTPhgxBfJcPxI29JzAqQjBdTxGWKG4XBzhpwqzuRCOwU_bFhThbEnvBBN3VzQvy5Fz4_DZ6Ep_khg',
-        },
-    }
-    const initBody = { meetUpId: id }
-    let closeBody = {
-        id: id,
-        category: detail.category,
-        title: detail.title,
-        contents: detail.contents,
-        maxPeople: detail.maxPeople,
-        isOpenYn: detail.isOpenYn,
-        isRecruiting: false,
-        leaderId: detail.leaderId,
-    }
-
-    const postMeetUp = (api, body) => {
-        axios
-            .post(`${uri}${api}`, body, config)
-            .then((res) => {
-                console.log(res)
-                res && setIsOpenPopUp(false)
-            })
-            .catch(function (error) {
-                console.log('meet-up', error)
-            })
-    }
-
-    useEffect(() => {
-        if (status === 'JOINED') {
-            setInfo({
-                footerText: '밋업 탈퇴하기',
-                popUpText: `탈퇴하기를 누르면 복구할 수가 없습니다.\n정말로 탈퇴하시겠습니까?`,
-                handleOk: () => postMeetUp('/meetup/service/leave', initBody),
-            })
-        } else if (status === 'UNRELATED') {
-            setInfo({
-                footerText: '참여하기',
-                popUpText: `참여하기를 누르면 단톡방으로 이동합니다\n밋업에 참여하시겠습니까?`,
-                handleOk: () => postMeetUp('/meetup/service/join', initBody),
-            })
-        } else if (status === 'OWNED') {
-            if (detail.isRecruiting) {
-                setInfo({
-                    footerText: '마감하기',
-                    popUpText: `밋업을 마감하시겠습니까?`,
-                    handleOk: () =>
-                        postMeetUp('/meetup/service/update', closeBody),
-                })
-            } else {
-                setInfo({
-                    footerText: '재모집하기 ',
-                    popUpText: `밋업을 재모집하시겠습니까?`,
-                    handleOk: () =>
-                        postMeetUp('/meetup/service/update', {
-                            ...closeBody,
-                            isRecruiting: true,
-                        }),
-                })
-            }
-        }
-    }, [status])
+    const { footerText, popUpText, handleOk } = useMeetUpDetail(
+        detail,
+        status,
+        setIsOpenPopUp
+    )
 
     return (
         <Main>
